{{cookiecutter.project_slug}}/requirements.txt
# Hackathon Project - Essential Dependencies Only

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation
pydantic>=2.5.2
pydantic-settings==2.1.0

# HTTP and networking
httpx==0.25.2
requests==2.31.0

# MongoDB driver
motor==3.3.2
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
numpy==1.26.2

# AI/LLM
google-generativeai==0.3.0

# Development tools (optional)
pytest==7.4.3
black==23.11.0